+++ packages/core/src/indexing/query-filter-converter.ts
@@ -1,7 +1,17 @@
 import { Knex } from 'knex'
-import { NonBooleanValueFilterType, ObjectFilter, QueryFilters } from './query-filter-parser.js'
+import {
+  getValueType,
+  NonBooleanValueFilterType,
+  ObjectFilter,
+  QueryFilters,
+  ValueFilterType,
+} from './query-filter-parser.js'
 
 export const DATA_FIELD = 'stream_content'
+
+export function contentKey(field: string): string {
+  return `${DATA_FIELD}->>'${field}'`
+}
 
 type DBQuery = Knex.QueryBuilder
 type WhereFunc = (DBQuery) => DBQuery
@@ -52,6 +62,24 @@ function handleQuery(
   }
 }
 
+function typeAsCast(tpe: ValueFilterType): string {
+  switch (tpe) {
+    case 'boolean':
+      return 'boolean'
+    default:
+      return nonBooleanTypeAsCast(tpe)
+  }
+}
+
+function nonBooleanTypeAsCast(tpe: NonBooleanValueFilterType): string {
+  switch (tpe) {
+    case 'number':
+      return 'numeric'
+    case 'string':
+      return 'varchar'
+  }
+}
+
 function handleIn<T extends number | string>(
   query: DBQuery,
   key: string,
@@ -62,10 +90,7 @@ function handleIn<T extends number | string>(
   combinator?: Combinator
 ): DBQuery {
   const arrValue = value.map((v) => v.toString()).join(',')
-  let cast = 'int'
-  if (tpe == 'string') {
-    cast = tpe
-  }
+  const cast = nonBooleanTypeAsCast(tpe)
   const inner = (bldr) => {
     let op = ' in '
     if (negated) {
@@ -90,7 +115,7 @@ function handleWhereQuery(state: ConversionState<ObjectFilter>): ConvertedQueryF
   for (const filterKey in state.filter) {
     select.push(filterKey)
     const value = state.filter[filterKey]
-    const key = `${DATA_FIELD}->>'${filterKey}'`
+    const key = contentKey(filterKey)
 
     switch (value.op) {
       case 'null': {
@@ -134,13 +159,8 @@ function handleWhereQuery(state: ConversionState<ObjectFilter>): ConvertedQueryF
       }
       default: {
         const isFirst = first
+        const cast = typeAsCast(getValueType(value.value))
         const old = where
-        let cast = 'boolean'
-        if (typeof value.value == 'number') {
-          cast = 'int'
-        } else if (typeof value.value == 'string') {
-          cast = 'varchar'
-        }
         where = (bldr) => {
           const b = old(bldr)
           let queryValue = value.value