+++ packages/stream-caip10-link-handler/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [2.24.0-rc.0](https://github.com/ceramicnetwork/js-ceramic/compare/@ceramicnetwork/stream-caip10-link-handler@2.23.0...@ceramicnetwork/stream-caip10-link-handler@2.24.0-rc.0) (2023-08-07)
+
+**Note:** Version bump only for package @ceramicnetwork/stream-caip10-link-handler
+
+
+
+
+
 # [2.23.0](https://github.com/ceramicnetwork/js-ceramic/compare/@ceramicnetwork/stream-caip10-link-handler@2.23.0-rc.1...@ceramicnetwork/stream-caip10-link-handler@2.23.0) (2023-07-31)
 
 **Note:** Version bump only for package @ceramicnetwork/stream-caip10-link-handler