+++ packages/stream-model-instance-handler/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.13.0-rc.0](https://github.com/ceramicnetwork/js-ceramic/compare/@ceramicnetwork/stream-model-instance-handler@1.12.0...@ceramicnetwork/stream-model-instance-handler@1.13.0-rc.0) (2023-08-07)
+
+**Note:** Version bump only for package @ceramicnetwork/stream-model-instance-handler
+
+
+
+
+
 # [1.12.0](https://github.com/ceramicnetwork/js-ceramic/compare/@ceramicnetwork/stream-model-instance-handler@1.12.0-rc.1...@ceramicnetwork/stream-model-instance-handler@1.12.0) (2023-07-31)
 
 **Note:** Version bump only for package @ceramicnetwork/stream-model-instance-handler